move updateEntity to provider
